requirements.txt
torch
torchvision
jupyterlab
scikit-learn
numpy
pandas
pytorch-ignite
matplotlib
tqdm

